refactored some test classes into the production folder for ease of testing
